+++ Lesson1/TriangleS.rb
@@ -1,7 +1,7 @@
 puts "Enter triangle's base, please"
-base = gets.chomp.to_i
+base = gets.chomp.to_f
 puts "Enter triangle's height, please"
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 Square = 0.5 * base * height
 puts "Square = #{Square}"
